mean-stack

+++ mean-stack/src/app/posts/post-list/post-list.component.ts
@@ -26,7 +26,9 @@ export class PostListComponent implements OnInit, OnDestroy {
   }
 
   onDelete(postId: string): any {
-    this.postsService.deletePost(postId);
+    if (confirm('Are you sure to delete it?')) {
+      this.postsService.deletePost(postId);
+    }
   }
 
   ngOnDestroy(): any {
